Improve error handling when link() fails
Use the normal convention of saving errno in a local variable right
after a system call has failed. (To be safe in case some other
irrelevant system call fails before we get to output the message for
errno.)

Use both Util::symbolicErrno() (so that a developer sees the actual
symbolic error name as in manpages and code) and strerror() (for
nominally helpful plaintext explanation of what that errno value
means) in the output.

Use explicit std:: prefix for strerror() (because that is cleaner and
it is our convention).

Use << operators in the LOG_FOO calls instead of concatenating strings
with the + operator.

Signed-off-by: Tor Lillqvist <tml@collabora.com>
Change-Id: I75f42051f81661eea62235019ef9a0925239ce00